Remove inclusion of election.hpp from other headers
Moves election_behavior enum and vote_with_weight_info classes in to their own files.
Before:
[118/118] Linking CXX executable rpc_test
ninja  1208.77s user 85.03s system 875% cpu 2:27.78 total
After:
[39/39] Linking CXX executable rpc_test
ninja  338.18s user 25.74s system 714% cpu 50.953 total
