+++ nano/node/active_transactions.hpp
@@ -1,8 +1,9 @@
 #pragma once
 
 #include <nano/lib/numbers.hpp>
-#include <nano/node/election.hpp>
+#include <nano/node/election_behavior.hpp>
 #include <nano/node/election_insertion_result.hpp>
+#include <nano/node/vote_with_weight_info.hpp>
 #include <nano/node/voting.hpp>
 #include <nano/secure/common.hpp>
