+++ nano/node/election_behavior.hpp
@@ -1,0 +1,22 @@
+#pragma once
+
+namespace nano
+{
+enum class election_behavior
+{
+	normal,
+	/**
+	 * Hinted elections:
+	 * - shorter timespan
+	 * - limited space inside AEC
+	 */
+	hinted,
+	/**
+	 * Optimistic elections:
+	 * - shorter timespan
+	 * - limited space inside AEC
+	 * - more frequent confirmation requests
+	 */
+	optimistic,
+};
+}
