+++ nano/node/vote_cache.cpp
@@ -1,4 +1,5 @@
 #include <nano/lib/tomlconfig.hpp>
+#include <nano/node/election.hpp>
 #include <nano/node/node.hpp>
 #include <nano/node/vote_cache.hpp>
